pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "litedocs"
version = "0.4.0"
authors = [
  { name="Litestartup", email="young@litestartup.com" },
]
description = "A lightweight, Markdown folder-based static documentation rendering tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.3",
    "markdown-it-py>=3.0.0",
    "mdit-py-plugins>=0.4.0",
    "python-frontmatter>=1.1.0",
    "typer>=0.12.0",
    "watchfiles>=0.21.0",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.scripts]
litedocs = "litedocs.cli:app"

[tool.setuptools.packages.find]
include = ["litedocs*"]

[tool.setuptools.package-data]
litedocs = [
    "themes/**/*.html",
    "themes/**/*.css",
    "themes/**/*.js",
    "themes/**/*.json",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
